# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ssomero API",
    "version": "1.0.0",
    "description": "OpenAPI description for Ssomero backend (students, courses, auth, etc.)."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        },
        "required": ["email","password"]
      },
      "RegisterUserDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" },
          "fullName": { "type": "string" }
        },
        "required": ["email","password"]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "refreshToken": { "type": "string" }
        },
        "required": ["token","refreshToken"]
      },
      "RefreshRequestDto": {
        "type": "object",
        "properties": {
          "refreshToken": { "type": "string" }
        },
        "required": ["refreshToken"]
      },
...
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "User login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequestDto" } }
          }
        },
        "responses": {
          "200": { "description": "Auth response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponseDto" } } } },
          "401": { "description": "Unauthorized" }
        },
        "security": []
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register user",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterUserDto" } } }
        },
        "responses": {
          "200": { "description": "Auth response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponseDto" } } } },
          "400": { "description": "Bad request" }
        },
        "security": []
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Refresh JWT using refresh token",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequestDto" } } }
        },
        "responses": {
          "200": { "description": "Auth response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponseDto" } } } },
          "401": { "description": "Unauthorized" }
        },
        "security": []
      }
    },
...
  }
}